modify. 유저 알림 수정 로직 변경 및 공통 hook 생성
