+++ src/components/templates/Admin/Admin.tsx
@@ -1,34 +1,16 @@
 import { Spacing } from "@components/common/Spacing";
 import { LoginHeader } from "@components/templates/Header";
 import { Info } from "@components/templates/Info";
-import { FirestoreError, doc, updateDoc } from "@firebase/firestore";
+import { useChangeToggle } from "@hooks/useChangeToggle";
 import { useOutletProps } from "@hooks/useOutletProps";
 import { useSignOut } from "@hooks/useSignOut";
-import { Alert } from "@utils/Alert";
-import { fb } from "@utils/firebase";
-import { Fragment, useCallback } from "react";
+import { Fragment } from "react";
 import { Outlet } from "react-router";
 
 export const AdminTemplate = () => {
   const { account, accountState, teamState } = useOutletProps();
   const { onSignOut } = useSignOut();
-
-  const onChangeNotification = useCallback(
-    (userDocId: string) => async (notification: boolean) => {
-      try {
-        const docRef = doc(fb.db, "accounts", userDocId);
-
-        await updateDoc(docRef, { isNotification: notification });
-
-        accountState.onChangeAccountNotify(notification);
-      } catch (e) {
-        if (e instanceof FirestoreError) {
-          Alert.error({ html: fb.errors(e.code) });
-        }
-      }
-    },
-    [accountState]
-  );
+  const { onChangeNotification } = useChangeToggle();
 
   return (
     <Fragment>
