+++ src/components/templates/User/User.tsx
@@ -3,33 +3,15 @@ import { LoginHeader } from "@components/templates/Header";
 import { Info } from "@components/templates/Info";
 import { UserInviteTemplate } from "@components/templates/User/Invitation";
 import { UserTeamTemplate } from "@components/templates/User/Team";
-import { FirestoreError, doc, updateDoc } from "@firebase/firestore";
+import { useChangeToggle } from "@hooks/useChangeToggle";
 import { useOutletProps } from "@hooks/useOutletProps";
 import { useSignOut } from "@hooks/useSignOut";
-import { Alert } from "@utils/Alert";
-import { fb } from "@utils/firebase";
-import { Fragment, useCallback } from "react";
+import { Fragment } from "react";
 
 export const UserTemplate = () => {
-  const { account, accountState, teamState } = useOutletProps();
+  const { account, teamState } = useOutletProps();
   const { onSignOut } = useSignOut();
-
-  const onChangeNotification = useCallback(
-    (userDocId: string) => async (notification: boolean) => {
-      try {
-        const docRef = doc(fb.db, "accounts", userDocId);
-
-        await updateDoc(docRef, { isNotification: notification });
-
-        accountState.onChangeAccountNotify(notification);
-      } catch (e) {
-        if (e instanceof FirestoreError) {
-          Alert.error({ html: fb.errors(e.code) });
-        }
-      }
-    },
-    [accountState]
-  );
+  const { onChangeNotification } = useChangeToggle();
 
   return (
     <Fragment>
